Refactor Sidekiq::RedisConnection
This minimize the amount of code in each adapter.
It also test both adapters with the same test suite (minus a few conditionals)
