+++ bin/sidekiqload
@@ -12,7 +12,7 @@ require_relative "../lib/sidekiq/cli"
 require_relative "../lib/sidekiq/launcher"
 
 if ENV["SIDEKIQ_REDIS_CLIENT"]
-  Sidekiq::RedisConnection.adapter = Sidekiq::RedisClientConnection
+  Sidekiq::RedisConnection.adapter = :redis_client
 end
 
 Sidekiq.configure_server do |config|
