+++ test/test_redis_connection.rb
@@ -6,8 +6,6 @@ require "sidekiq/cli"
 describe Sidekiq::RedisConnection do
   describe "create" do
     before do
-      @previous_adapter = Sidekiq::RedisConnection.adapter
-      Sidekiq::RedisConnection.adapter = Sidekiq::RedisConnection
       Sidekiq.options = Sidekiq::DEFAULTS.dup
       @old = ENV["REDIS_URL"]
       ENV["REDIS_URL"] = "redis://localhost/15"
@@ -15,7 +13,6 @@ describe Sidekiq::RedisConnection do
 
     after do
       ENV["REDIS_URL"] = @old
-      Sidekiq::RedisConnection.adapter = @previous_adapter
     end
 
     # To support both redis-rb 3.3.x #client and 4.0.x #_client
@@ -27,9 +24,32 @@ describe Sidekiq::RedisConnection do
       end
     end
 
+    def config_for(redis)
+      client = client_for(redis)
+      if client.respond_to?(:config)
+        client.config
+      else
+        client
+      end
+    end
+
+    def self.redis_client?
+      Sidekiq::RedisConnection.adapter == Sidekiq::RedisConnection::RedisClientAdapter
+    end
+
+    if redis_client?
+      def client_class
+        Sidekiq::RedisConnection::RedisClientAdapter::CompatClient
+      end
+    else
+      def client_class
+        Redis
+      end
+    end
+
     it "creates a pooled redis connection" do
       pool = Sidekiq::RedisConnection.create
-      assert_equal Redis, pool.checkout.class
+      assert_equal client_class, pool.checkout.class
     end
 
     # Readers for these ivars should be available in the next release of
@@ -90,8 +110,13 @@ describe Sidekiq::RedisConnection do
 
     it "disables client setname with nil id" do
       pool = Sidekiq::RedisConnection.create(id: nil)
-      assert_equal Redis, pool.checkout.class
-      assert_equal "redis://localhost:6379/15", pool.checkout.connection.fetch(:id)
+      assert_equal client_class, pool.checkout.class
+      client = client_for(pool.checkout)
+      if self.class.redis_client?
+        assert_nil client.id
+      else
+        assert_equal "redis://localhost:6379/15", client.id
+      end
     end
 
     describe "network_timeout" do
@@ -99,48 +124,65 @@ describe Sidekiq::RedisConnection do
         pool = Sidekiq::RedisConnection.create(network_timeout: 8)
         redis = pool.checkout
 
-        assert_equal 8, client_for(redis).timeout
+        assert_equal 8, client_for(redis).read_timeout
       end
 
       it "uses the default network_timeout if none specified" do
         pool = Sidekiq::RedisConnection.create
         redis = pool.checkout
 
-        assert_equal 5, client_for(redis).timeout
+        if self.class.redis_client?
+          assert_equal 1.0, client_for(redis).read_timeout
+        else
+          assert_equal 5, client_for(redis).read_timeout
+        end
       end
     end
 
     describe "namespace" do
-      it "uses a given :namespace set by a symbol key" do
-        pool = Sidekiq::RedisConnection.create(namespace: "xxx")
-        assert_equal "xxx", pool.checkout.namespace
-      end
+      if redis_client?
+        it "isn't supported" do
+          Kernel.stub(:exit, ->(code) { raise "Exited #{code}" }) do
+            output = capture_logging do
+              error = assert_raises RuntimeError do
+                Sidekiq::RedisConnection.create(namespace: "xxx")
+              end
+              assert_includes error.message, "Exited -127"
+            end
+            assert_includes output, "Your Redis configuration uses the namespace 'xxx' but this feature isn't supported by redis-client"
+          end
+        end
+      else
+        it "uses a given :namespace set by a symbol key" do
+          pool = Sidekiq::RedisConnection.create(namespace: "xxx")
+          assert_equal "xxx", pool.checkout.namespace
+        end
 
-      it "uses a given :namespace set by a string key" do
-        pool = Sidekiq::RedisConnection.create("namespace" => "xxx")
-        assert_equal "xxx", pool.checkout.namespace
-      end
+        it "uses a given :namespace set by a string key" do
+          pool = Sidekiq::RedisConnection.create("namespace" => "xxx")
+          assert_equal "xxx", pool.checkout.namespace
+        end
 
-      it "uses given :namespace over :namespace from Sidekiq.options" do
-        Sidekiq.options[:namespace] = "xxx"
-        pool = Sidekiq::RedisConnection.create(namespace: "yyy")
-        assert_equal "yyy", pool.checkout.namespace
+        it "uses given :namespace over :namespace from Sidekiq.options" do
+          Sidekiq.options[:namespace] = "xxx"
+          pool = Sidekiq::RedisConnection.create(namespace: "yyy")
+          assert_equal "yyy", pool.checkout.namespace
+        end
       end
     end
 
     describe "socket path" do
       it "uses a given :path" do
         pool = Sidekiq::RedisConnection.create(path: "/var/run/redis.sock")
-        assert_equal "unix", client_for(pool.checkout).scheme
-        assert_equal "/var/run/redis.sock", pool.checkout.connection.fetch(:location)
-        assert_equal 15, pool.checkout.connection.fetch(:db)
+        config = config_for(pool.checkout)
+        assert_equal "/var/run/redis.sock", config.path
       end
 
       it "uses a given :path and :db" do
         pool = Sidekiq::RedisConnection.create(path: "/var/run/redis.sock", db: 8)
-        assert_equal "unix", client_for(pool.checkout).scheme
-        assert_equal "/var/run/redis.sock", pool.checkout.connection.fetch(:location)
-        assert_equal 8, pool.checkout.connection.fetch(:db)
+        config = config_for(pool.checkout)
+        assert_equal "/var/run/redis.sock", config.path
+        assert_equal 8, config.db
       end
     end
 
@@ -159,31 +201,47 @@ describe Sidekiq::RedisConnection do
     end
 
     describe "driver" do
-      it "uses redis' ruby driver" do
-        pool = Sidekiq::RedisConnection.create
-        redis = pool.checkout
+      if redis_client?
+        it "uses ruby driver by default" do
+          pool = Sidekiq::RedisConnection.create
+          config = config_for(pool.checkout)
 
-        assert_equal Redis::Connection::Ruby, redis.instance_variable_get(:@client).driver
-      end
+          assert_equal RedisClient::Connection, config.driver
+        end
 
-      it "uses redis' default driver if there are many available" do
-        redis_driver = Object.new
-        Redis::Connection.drivers << redis_driver
+        it "uses hiredis driver if asked" do
+          pool = Sidekiq::RedisConnection.create(driver: "hiredis")
+          config = config_for(pool.checkout)
 
-        pool = Sidekiq::RedisConnection.create
-        redis = pool.checkout
+          assert_equal RedisClient::HiredisConnection, config.driver
+        end
+      else
+        it "uses redis' ruby driver" do
+          pool = Sidekiq::RedisConnection.create
+          redis = pool.checkout
 
-        assert_equal redis_driver, redis.instance_variable_get(:@client).driver
-      ensure
-        Redis::Connection.drivers.pop
-      end
+          assert_equal Redis::Connection::Ruby, redis.instance_variable_get(:@client).driver
+        end
 
-      it "uses a given :driver" do
-        redis_driver = Object.new
-        pool = Sidekiq::RedisConnection.create(driver: redis_driver)
-        redis = pool.checkout
+        it "uses redis' default driver if there are many available" do
+          redis_driver = Object.new
+          Redis::Connection.drivers << redis_driver
 
-        assert_equal redis_driver, redis.instance_variable_get(:@client).driver
+          pool = Sidekiq::RedisConnection.create
+          redis = pool.checkout
+
+          assert_equal redis_driver, redis.instance_variable_get(:@client).driver
+        ensure
+          Redis::Connection.drivers.pop
+        end
+
+        it "uses a given :driver" do
+          redis_driver = Object.new
+          pool = Sidekiq::RedisConnection.create(driver: redis_driver)
+          redis = pool.checkout
+
+          assert_equal redis_driver, redis.instance_variable_get(:@client).driver
+        end
       end
     end
